Update slice.js to duck patterns
