+++ src/slice.js
@@ -1,4 +1,16 @@
 import { createSlice } from '@reduxjs/toolkit';
+
+import {
+  fetchRegions,
+  fetchCategories,
+  fetchRestaurants,
+  fetchRestaurant,
+  postLogin,
+  postReview,
+} from './services/api';
+
+import { saveItem } from './services/storage';
+
 import { equal } from './utils';
 
 const initialReviewFields = {
@@ -126,4 +138,91 @@ const { actions, reducer } = createSlice({
   },
 });
 
-export { actions, reducer };
+export const {
+  setRegions,
+  setCategories,
+  setRestaurants,
+  setRestaurant,
+  setAccessToken,
+  clearReviewFields,
+  setReviews,
+  selectRegion,
+  selectCategory,
+  changeLoginField,
+  changeReviewField,
+  logout,
+} = actions;
+
+export function loadInitialData() {
+  return async (dispatch) => {
+    const regions = await fetchRegions();
+    dispatch(setRegions(regions));
+
+    const categories = await fetchCategories();
+    dispatch(setCategories(categories));
+  };
+}
+
+export function loadRestaurants() {
+  return async (dispatch, getState) => {
+    const {
+      selectedRegion: region,
+      selectedCategory: category,
+    } = getState();
+
+    if (!region || !category) {
+      return;
+    }
+
+    const restaurants = await fetchRestaurants({
+      regionName: region.name,
+      categoryId: category.id,
+    });
+    dispatch(setRestaurants(restaurants));
+  };
+}
+
+export function loadRestaurant({ restaurantId }) {
+  return async (dispatch) => {
+    dispatch(setRestaurant(null));
+
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setRestaurant(restaurant));
+  };
+}
+
+export function requestLogin() {
+  return async (dispatch, getState) => {
+    const { loginFields: { email, password } } = getState();
+
+    const accessToken = await postLogin({ email, password });
+
+    saveItem('accessToken', accessToken);
+
+    dispatch(setAccessToken(accessToken));
+  };
+}
+
+export function loadReview({ restaurantId }) {
+  return async (dispatch) => {
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setReviews(restaurant.reviews));
+  };
+}
+
+export function sendReview({ restaurantId }) {
+  return async (dispatch, getState) => {
+    const { accessToken, reviewFields: { score, description } } = getState();
+
+    await postReview({
+      accessToken, restaurantId, score, description,
+    });
+
+    dispatch(loadReview({ restaurantId }));
+    dispatch(clearReviewFields());
+  };
+}
+
+export default reducer;
