+++ src/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import reducer from './reducer';
+import reducer from './slice';
 
 const store = configureStore({
   reducer,
